refactor: use domain types only in domain layer
In `internal/jimm/juju/model.go` we provided a function to convert from Juju's RPC types to our internal type. With the introduction of TargetController as a JAAS specific field we can see that this breaks down and the API types should not be used at all in the domain layer.
This commit moves the helper function up to the API layer and ensures tests use the domain type.
We also revert previous changes to the Juju import.
