+++ internal/jujuapi/types_test.go
@@ -1,0 +1,114 @@
+// Copyright 2025 Canonical.
+
+package jujuapi
+
+import (
+	"testing"
+
+	qt "github.com/frankban/quicktest"
+	"github.com/google/go-cmp/cmp"
+	jujuparams "github.com/juju/juju/rpc/params"
+	"github.com/juju/names/v5"
+
+	"github.com/canonical/jimm/v3/internal/jimm/juju"
+)
+
+func TestModelCreateArgs(t *testing.T) {
+	c := qt.New(t)
+
+	tests := []struct {
+		about         string
+		args          jujuparams.ModelCreateArgs
+		expectedArgs  *juju.ModelCreateArgs
+		expectedError string
+	}{{
+		about: "all ok",
+		args: jujuparams.ModelCreateArgs{
+			Name:               "test-model",
+			OwnerTag:           names.NewUserTag("alice@canonical.com").String(),
+			CloudTag:           names.NewCloudTag("test-cloud").String(),
+			CloudCredentialTag: names.NewCloudCredentialTag("test-cloud/alice@canonical.com/test-credential-1").String(),
+		},
+		expectedArgs: &juju.ModelCreateArgs{
+			Name:            "test-model",
+			Owner:           names.NewUserTag("alice@canonical.com"),
+			Cloud:           names.NewCloudTag("test-cloud"),
+			CloudCredential: names.NewCloudCredentialTag("test-cloud/alice@canonical.com/test-credential-1"),
+		},
+	}, {
+		about: "name not specified",
+		args: jujuparams.ModelCreateArgs{
+			OwnerTag:           names.NewUserTag("alice@canonical.com").String(),
+			CloudTag:           names.NewCloudTag("test-cloud").String(),
+			CloudCredentialTag: names.NewCloudCredentialTag("test-cloud/alice/test-credential-1").String(),
+		},
+		expectedError: "name not specified",
+	}, {
+		about: "invalid owner tag",
+		args: jujuparams.ModelCreateArgs{
+			Name:               "test-model",
+			OwnerTag:           "alice@canonical.com",
+			CloudTag:           names.NewCloudTag("test-cloud").String(),
+			CloudCredentialTag: names.NewCloudCredentialTag("test-cloud/alice/test-credential-1").String(),
+		},
+		expectedError: `"alice@canonical.com" is not a valid tag`,
+	}, {
+		about: "invalid cloud tag",
+		args: jujuparams.ModelCreateArgs{
+			Name:               "test-model",
+			OwnerTag:           names.NewUserTag("alice@canonical.com").String(),
+			CloudTag:           "test-cloud",
+			CloudCredentialTag: names.NewCloudCredentialTag("test-cloud/alice/test-credential-1").String(),
+		},
+		expectedError: `"test-cloud" is not a valid tag`,
+	}, {
+		about: "invalid cloud credential tag",
+		args: jujuparams.ModelCreateArgs{
+			Name:               "test-model",
+			OwnerTag:           names.NewUserTag("alice@canonical.com").String(),
+			CloudTag:           names.NewCloudTag("test-cloud").String(),
+			CloudCredentialTag: "test-credential-1",
+		},
+		expectedError: "invalid cloud credential tag",
+	}, {
+		about: "cloud does not match cloud credential cloud",
+		args: jujuparams.ModelCreateArgs{
+			Name:               "test-model",
+			OwnerTag:           names.NewUserTag("alice@canonical.com").String(),
+			CloudTag:           names.NewCloudTag("test-cloud").String(),
+			CloudCredentialTag: names.NewCloudCredentialTag("another-cloud/alice/test-credential-1").String(),
+		},
+		expectedError: "cloud credential cloud mismatch",
+	}, {
+		about: "owner tag not specified",
+		args: jujuparams.ModelCreateArgs{
+			Name:               "test-model",
+			CloudTag:           names.NewCloudTag("test-cloud").String(),
+			CloudCredentialTag: names.NewCloudCredentialTag("test-cloud/alice/test-credential-1").String(),
+		},
+		expectedError: "owner tag not specified",
+	}}
+
+	opts := []cmp.Option{
+		cmp.Comparer(func(t1, t2 names.UserTag) bool {
+			return t1.String() == t2.String()
+		}),
+		cmp.Comparer(func(t1, t2 names.CloudTag) bool {
+			return t1.String() == t2.String()
+		}),
+		cmp.Comparer(func(t1, t2 names.CloudCredentialTag) bool {
+			return t1.String() == t2.String()
+		}),
+	}
+	for _, test := range tests {
+		c.Run(test.about, func(c *qt.C) {
+			a, err := toAddModelArgs(test.args)
+			if test.expectedError == "" {
+				c.Assert(err, qt.IsNil)
+				c.Assert(a, qt.CmpEquals(opts...), test.expectedArgs)
+			} else {
+				c.Assert(err, qt.ErrorMatches, test.expectedError)
+			}
+		})
+	}
+}
